Change Google function to use alternate Firefox profile
When using i3wm I want to be able to Google things in vim and have them
open in a browser in my current workspace. However, if I already have an
instance of Firefox running in a different workspace the new tab will
open there. To get around this I created a second Firefox profile called
alt which I now use in the Google function.